Fix AO/CO visimap first row number integer overflow (#4564)
There is a scenario during AO/CO delete or update where the first row number
obtained is negative. The error is caused when the first row number in the
aovisimap of an AO/CO table exceeds int max. There is currently a bug where
the first row number is retrieved from the tuple as an int, but the first row
number is an int64. We fixed this by retrieving as an int64.

+++ src/backend/access/appendonly/appendonly_visimap_entry.c
@@ -386,7 +386,7 @@ AppendOnlyVisimapEntry_CoversTuple(
 		AppendOnlyVisimapEntry *visiMapEntry,
 		AOTupleId *tupleId)
 {
-	int rowNum;
+	int64 rowNum;
 
 	Assert(visiMapEntry);
 	Assert(tupleId);
@@ -414,8 +414,8 @@ AppendOnlyVisimapEntry_GetFirstRowNum(
 		AppendOnlyVisimapEntry *visiMapEntry,
 		AOTupleId *tupleId)
 {
-	(void)visiMapEntry;
-	int rowNum;
+	(void) visiMapEntry;
+	int64 rowNum;
 
 	rowNum = AOTupleIdGet_rowNum(tupleId);
 	return (rowNum / APPENDONLY_VISIMAP_MAX_RANGE) * APPENDONLY_VISIMAP_MAX_RANGE;
